Conflicts:
	cmd-attach-session.c
	cmd-choose-buffer.c
	cmd-detach-client.c
	cmd-if-shell.c
	cmd-kill-session.c
	cmd-new-session.c
	cmd-new-window.c
	cmd-respawn-pane.c
	cmd-respawn-window.c
	cmd-run-shell.c
	cmd-set-environment.c
	cmd-set-option.c
	cmd-show-environment.c
	cmd-show-messages.c
	cmd-show-options.c
	cmd-split-window.c
	cmd.c
	tmux.c
	tmux.h

+++ server.c
@@ -172,9 +172,12 @@ server_start(struct event_base *base, int lockfd, char *lockfile)
 	}
 	close(pair[0]);
 
+	if (log_get_level() > 3)
+		tty_create_log();
+
 #ifdef __OpenBSD__
-	if (pledge("stdio rpath wpath cpath fattr unix recvfd proc exec tty "
-	    "ps", NULL) != 0)
+	if (pledge("stdio rpath wpath cpath fattr unix getpw recvfd proc exec "
+	    "tty ps", NULL) != 0)
 		fatal("pledge failed");
 #endif
 
@@ -185,9 +188,8 @@ server_start(struct event_base *base, int lockfd, char *lockfile)
 	TAILQ_INIT(&session_groups);
 	mode_key_init_trees();
 	key_bindings_init();
-	utf8_build();
 
-	start_time = time(NULL);
+	gettimeofday(&start_time, NULL);
 
 	server_fd = server_create_socket();
 	if (server_fd == -1)
@@ -195,9 +197,11 @@ server_start(struct event_base *base, int lockfd, char *lockfile)
 	server_update_socket();
 	server_client_create(pair[1]);
 
-	unlink(lockfile);
-	free(lockfile);
-	close(lockfd);
+	if (lockfd >= 0) {
+		unlink(lockfile);
+		free(lockfile);
+		close(lockfd);
+	}
 
 	start_cfg();
 
@@ -298,7 +302,7 @@ server_update_socket(void)
 
 /* Callback for server socket. */
 void
-server_accept(int fd, short events, unused void *data)
+server_accept(int fd, short events, __unused void *data)
 {
 	struct sockaddr_storage	sa;
 	socklen_t		slen = sizeof sa;
